requirements.txt
# Core dependencies
requests>=2.28.0
python-dotenv>=1.0.0
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
pillow>=10.0.0
nltk>=3.8.0
rich>=13.0.0
seaborn>=0.12.0

# AI Models
transformers>=4.38.0
torch>=2.2.0
datasets>=2.16.0
huggingface-hub>=0.20.0

# API Framework
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.9
pydantic>=2.0.0

# Development & Debugging
jupyter>=1.0.0
ipykernel>=6.0.0
debugpy>=1.8.0
psutil>=5.9.0

# Type stubs
types-requests==2.31.0.20240311
types-psutil>=5.9.5
types-Pillow>=10.0.0

# Testing dependencies
pytest>=7.3.1
pytest-asyncio>=0.21.0
httpx>=0.24.0  # For TestClient in FastAPI
unittest-mock>=1.3.0
